feat: add support for native images to the rockcraft-gradle plugin (#132)
This feature lets users create application rocks with native image of
the application, compiled using GraalVM native image. It also includes
support to create a build-rock that acts as a build and test container
to build native images of applications.

This PR adds the native-image support to the gradle front-end, while
reusing most of the common backend added by #115.